eka_backend/requirements.txt

Django>=4.2,<5.0


djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4


django-cors-headers>=4.3,<4.4


psycopg2-binary>=2.9,<2.10


gunicorn>=21.2,<22.0